backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
google-auth==2.28.1
requests==2.32.5
email-validator
python-multipart
